requirements.txt
streamlit
torch
torchvision
pytorch-gradcam
opencv-python-headless
pillow
numpy
plotly
